# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: NS Scrap ERP Next API
  version: 0.1.0
  description: >
    Baseline API catalog for the active Next.js app. This file records current
    route handlers first; detailed request and response schemas should be added
    as each module batch hardens its contract.
servers:
  - url: http://localhost:3000
    description: Local Next.js development server
  - url: https://new-ns-scrap-erp.vercel.app
    description: Current Vercel deployment, if available
security:
  - SupabaseSession: []
...
paths:
...
  /api/stock/balance:
    get:
      tags: [Stock]
      operationId: getStockBalance
      summary: Read stock balance aggregated from stock ledger
      parameters:
        - $ref: "#/components/parameters/StockQ"
        - $ref: "#/components/parameters/StockProductId"
        - $ref: "#/components/parameters/StockBranchId"
        - $ref: "#/components/parameters/StockWarehouseId"
        - $ref: "#/components/parameters/StockStatus"
        - $ref: "#/components/parameters/StockLotNo"
        - $ref: "#/components/parameters/StockAsOf"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Format"
      responses:
        "200":
          description: Aggregated stock balance grouped by product, metal group, branch, warehouse, status, and lot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockBalanceResponse"
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    SupabaseSession:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Supabase Auth session cookie or equivalent authenticated request context.
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 10
        maximum: 500
        default: 50
    Format:
      name: format
      in: query
      schema:
        type: string
        enum: [json, xlsx]
        default: json
...
    StockAsOf:
      name: asOf
      in: query
      schema:
        type: string
        format: date
    StockBranchId:
      name: branchId
      in: query
      schema:
        type: string
...
    StockLotNo:
      name: lotNo
      in: query
      schema:
        type: string
        maxLength: 80
...
    StockProductId:
      name: productId
      in: query
      schema:
        type: string
    StockQ:
      name: q
      in: query
      schema:
        type: string
        maxLength: 120
...
    StockStatus:
      name: status
      in: query
      schema:
        $ref: "#/components/schemas/StockStatus"
...
    StockWarehouseId:
      name: warehouseId
      in: query
      schema:
        type: string
...
  responses:
...
    Unauthorized:
      description: Missing or invalid Supabase session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      additionalProperties: true
...
    StockBalanceResponse:
      type: object
      required: [byStatus, page, pageSize, rows, summary, total]
      properties:
        byStatus:
          type: array
          items:
            type: object
            properties:
              count:
                type: integer
              qty:
                type: number
              status:
                $ref: "#/components/schemas/StockStatus"
              value:
                type: number
        page:
          type: integer
        pageSize:
          type: integer
        rows:
          type: array
          items:
            $ref: "#/components/schemas/StockBalanceRow"
        summary:
          type: object
          additionalProperties: true
        total:
          type: integer
    StockBalanceRow:
      type: object
      properties:
        avgCost:
          type: number
        branchId:
          type: [string, "null"]
        branchName:
          type: [string, "null"]
        lastMovementDate:
          type: [string, "null"]
        lotNo:
          type: [string, "null"]
        notAvailableForSale:
          type: boolean
        productCode:
          type: [string, "null"]
        productId:
          type: string
        productMetalGroup:
          type: string
        productName:
          type: [string, "null"]
        qty:
          type: number
        status:
          $ref: "#/components/schemas/StockStatus"
        value:
          type: number
        warehouseId:
          type: [string, "null"]
        warehouseName:
          type: [string, "null"]
...
    StockStatus:
      type: string
      enum: [RM, WIP, FG]
...